chore: add sign component to company creation
