+++ src/components/Upload.tsx
@@ -31,6 +31,28 @@ function UploadedFile({
   );
 }
 
+export function UploadedFilePreview({
+  name,
+  size,
+}: {
+  name: string;
+  size: number;
+}) {
+  return (
+    <div className="flex items-center justify-between p-4 bg-elevation-200 border border-divider-50 rounded-lg">
+      <div className="flex gap-1 items-center">
+        <span className="max-w-32 text-text-300 text-sm font-medium leading-5 truncate">
+          {name}
+        </span>
+        <span className="text-text-200 text-xs font-normal leading-[18px]">
+          {size} KB
+        </span>
+        <CircleCheckIcon className="text-signal-success h-4 w-4 stroke-1" />
+      </div>
+    </div>
+  );
+}
+
 export default function Upload({
   label,
   description,
