+++ src/i18n/source.json
@@ -280,6 +280,7 @@
   "company.create.requiredInformation.description": "Please enter the required data to create a new company identity",
   "company.create.requiredInformation.title": "Company identity",
   "company.create.sign": "Sign",
+  "company.create.sign.confirm": "Confirm",
   "company.create.success.description": "We have successfully created your company and can add other signers",
   "company.create.success.finish": "Finish",
   "company.create.success.title": "Success",
@@ -290,6 +291,7 @@
   "company.name": "Legal company name",
   "company.preview.description": "Please check your company data before signing",
   "company.preview.title": "Preview company",
+  "company.proof_of_registration": "Registration document",
   "company.registration_date": "Registration date",
   "company.registration_number": "Registration number",
   "company.signer.add": "Add authorized signer",
@@ -361,6 +363,7 @@
   "contacts.create.requiredInformation.nodeId": "Node ID",
   "contacts.create.requiredInformation.personalName": "Legal full name",
   "contacts.create.requiredInformation.title": "Required",
+  "contacts.create.sign.cancel": "Cancel",
   "contacts.create.switchToCompany": "Company",
   "contacts.create.switchToMint": "Mint",
   "contacts.create.switchToPerson": "Person",
